requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
click==8.1.7
fastapi==0.115.5
Flask==3.1.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
pydantic==2.10.1
pydantic_core==2.27.1
pytesseract==0.3.13
python-multipart==0.0.17
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
Werkzeug==3.1.3
